requirements.txt
Flask==2.3.3
torch>=2.2
torchvision>=0.17
numpy==1.24.3
pandas==2.0.3
librosa==0.10.1
soundfile==0.12.1
matplotlib==3.7.2
scikit-learn==1.3.0
werkzeug>=3.0
Pillow==10.0.0
reportlab==4.4.3
xlsxwriter==3.2.5